+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- fix: SwiftUI view modifiers .postHogMask() and .postHogNoMask() on iOS 26 ([#414](https://github.com/PostHog/posthog-ios/pull/414))
+- fix: SwiftUI view modifiers .postHogMask() and .postHogNoMask() on iOS 26 ([#415](https://github.com/PostHog/posthog-ios/pull/415))
 
 ## 3.36.1 - 2025-12-16